Add macro to disable `std::vector` wrapper ctor.
This constructor allows direct construction of a 1-d `marray` from a
`std::vector`, but precludes using a `std::vector<int>` or similar to
specify the tensor lengths when using a constructor like:

```
std::vector<int> lengths{3,4,5};
Marray::marray(lengths); // Perhaps the user wanted a 3-d marray?
```

Importantly, the "old" MArray always used the interpretation as a
vector of lengths. This macros disables the `std::vector` constructor
and restores the old behavior.